refactor(issue-237): mover los eventos privados a PrivateEvents
ProgramacionTurnoDiarioSolicitada, DetalleTurno, DetalleFranjaOrdinaria y
DetalleSubFranja pasan a PrivateEvents.Programacion. Primer uso real del
grafo: el evento privado consume InformacionEmpleado desde PublicEvents.
