+++ src/Bitakora.ControlAsistencia.ControlHoras/Entities/ControlFranja.cs
@@ -1,5 +1,6 @@
 using Bitakora.ControlAsistencia.ControlHoras.ValueObjects;
 using Bitakora.ControlAsistencia.Contracts.Programacion.ValueObjects;
+using Bitakora.ControlAsistencia.PrivateEvents.Programacion;
 
 namespace Bitakora.ControlAsistencia.ControlHoras.Entities;
