+++ src/Bitakora.ControlAsistencia.ControlHoras/ValueObjects/DesgloseFranja.cs
@@ -1,4 +1,5 @@
 using Bitakora.ControlAsistencia.Contracts.Programacion.ValueObjects;
+using Bitakora.ControlAsistencia.PrivateEvents.Programacion;
 
 namespace Bitakora.ControlAsistencia.ControlHoras.ValueObjects;
