+++ src/Bitakora.ControlAsistencia.PrivateEvents/Programacion/DetalleSubFranja.cs
@@ -1,4 +1,4 @@
-namespace Bitakora.ControlAsistencia.Contracts.Programacion.ValueObjects;
+namespace Bitakora.ControlAsistencia.PrivateEvents.Programacion;
 
 /// <summary>
 /// Representacion plana de una sub-franja (descanso o extra) que viaja en eventos entre dominios.
