+++ src/Bitakora.ControlAsistencia.Programacion/SolicitarProgramacionTurnoFunction/Eventos/ProgramacionTurnoSolicitada.cs
@@ -1,4 +1,5 @@
 using Bitakora.ControlAsistencia.Contracts.Programacion.ValueObjects;
+using Bitakora.ControlAsistencia.PrivateEvents.Programacion;
 using Bitakora.ControlAsistencia.PublicEvents.Empleados;
 
 namespace Bitakora.ControlAsistencia.Programacion.SolicitarProgramacionTurnoFunction.Eventos;
