+++ tests/Bitakora.ControlAsistencia.ControlHoras.Tests/Entities/ClasificadorTrabajoTests.cs
@@ -8,6 +8,7 @@ using AwesomeAssertions;
 using Bitakora.ControlAsistencia.ControlHoras.ValueObjects;
 using Bitakora.ControlAsistencia.Contracts.Programacion.ValueObjects;
 using Bitakora.ControlAsistencia.ControlHoras.Entities;
+using Bitakora.ControlAsistencia.PrivateEvents.Programacion;
 
 namespace Bitakora.ControlAsistencia.ControlHoras.Tests.Entities;
