+++ tests/Bitakora.ControlAsistencia.ControlHoras.Tests/Entities/ConsolidadorDesgloseHorasTestData.cs
@@ -16,6 +16,7 @@
 
 using Bitakora.ControlAsistencia.ControlHoras.ValueObjects;
 using Bitakora.ControlAsistencia.Contracts.Programacion.ValueObjects;
+using Bitakora.ControlAsistencia.PrivateEvents.Programacion;
 
 namespace Bitakora.ControlAsistencia.ControlHoras.Tests.Entities;
