+++ tests/Bitakora.ControlAsistencia.ControlHoras.Tests/Entities/ControlFranjaTests.cs
@@ -3,6 +3,7 @@
 using AwesomeAssertions;
 using Bitakora.ControlAsistencia.Contracts.Programacion.ValueObjects;
 using Bitakora.ControlAsistencia.ControlHoras.Entities;
+using Bitakora.ControlAsistencia.PrivateEvents.Programacion;
 
 namespace Bitakora.ControlAsistencia.ControlHoras.Tests.Entities;
