+++ tests/Bitakora.ControlAsistencia.ControlHoras.Tests/Infraestructura/DesgloseHorasSerializacionTests.cs
@@ -8,6 +8,7 @@ using AwesomeAssertions;
 using Bitakora.ControlAsistencia.ControlHoras.ValueObjects;
 using Bitakora.ControlAsistencia.Contracts.Programacion.ValueObjects;
 using Bitakora.ControlAsistencia.ControlHoras.Infraestructura;
+using Bitakora.ControlAsistencia.PrivateEvents.Programacion;
 
 namespace Bitakora.ControlAsistencia.ControlHoras.Tests.Infraestructura;
