+++ tests/Bitakora.ControlAsistencia.ControlHoras.Tests/ValueObjects/DesgloseFranjaTests.cs
@@ -4,6 +4,7 @@
 using AwesomeAssertions;
 using Bitakora.ControlAsistencia.ControlHoras.ValueObjects;
 using Bitakora.ControlAsistencia.Contracts.Programacion.ValueObjects;
+using Bitakora.ControlAsistencia.PrivateEvents.Programacion;
 
 namespace Bitakora.ControlAsistencia.ControlHoras.Tests.ValueObjects;
