+++ tests/Bitakora.ControlAsistencia.ControlHoras.Tests/ValueObjects/DesgloseHorasDiscriminarTests.cs
@@ -22,6 +22,7 @@
 using AwesomeAssertions;
 using Bitakora.ControlAsistencia.ControlHoras.ValueObjects;
 using Bitakora.ControlAsistencia.Contracts.Programacion.ValueObjects;
+using Bitakora.ControlAsistencia.PrivateEvents.Programacion;
 
 namespace Bitakora.ControlAsistencia.ControlHoras.Tests.ValueObjects;
