+++ tests/Bitakora.ControlAsistencia.ControlHoras.Tests/ValueObjects/DesgloseHorasTests.cs
@@ -5,6 +5,7 @@
 using AwesomeAssertions;
 using Bitakora.ControlAsistencia.ControlHoras.ValueObjects;
 using Bitakora.ControlAsistencia.Contracts.Programacion.ValueObjects;
+using Bitakora.ControlAsistencia.PrivateEvents.Programacion;
 
 namespace Bitakora.ControlAsistencia.ControlHoras.Tests.ValueObjects;
